move logic to context.ts
